requirements.txt
annotated-types==0.7.0
anyio==4.6.0
exceptiongroup==1.2.2
fastapi==0.115.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.1
click==8.1.7
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
idna==3.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mypy-extensions==1.0.0
ollama==0.3.3
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
uvicorn==0.31.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mypy-extensions==1.0.0
ollama==0.3.3
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
uvicorn==0.31.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
mypy-extensions==1.0.0
numpy==2.1.3
ollama==0.3.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
cssutils==2.11.1
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
lxml==5.3.0
more-itertools==10.5.0
mypy-extensions==1.0.0
numpy==2.1.3
ollama==0.3.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
premailer==3.10.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
yagmail==0.15.293
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
more-itertools==10.5.0
mypy-extensions==1.0.0
numpy==2.1.3
ollama==0.3.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
premailer==3.10.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
yagmail==0.15.293
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
black==24.10.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
cssutils==2.11.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==30.5.0
fastapi==0.115.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
more-itertools==10.5.0
mypy-extensions==1.0.0
numpy==2.1.3
ollama==0.3.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
premailer==3.10.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
yagmail==0.15.293
